refactor: restructure CLI to cobra-style architecture
- Move CLI components to dedicated cli/ directory structure
- Separate scan and graph commands into individual modules
- Create clean main.rs entry point via RootCommand
- Organize args, commands, and root command logically
- Maintain full backward compatibility for existing functionality

New structure:
├── src/cli/
│   ├── mod.rs (public interface)
│   ├── root.rs (main command router)
│   ├── args.rs (argument definitions)
│   └── commands/
│       ├── scan.rs (security scanning)
│       └── graph.rs (call graph generation)

This follows Go cobra library patterns for better maintainability
and extensibility of CLI commands.
