+++ src/cli/commands/mod.rs
@@ -1,0 +1,5 @@
+pub mod scan;
+pub mod graph;
+
+pub use scan::run_scan_command;
+pub use graph::run_graph_command;
