+++ src/cli/mod.rs
@@ -1,0 +1,6 @@
+pub mod args;
+pub mod commands;
+pub mod root;
+
+pub use args::{Args, Commands, ScanArgs, GraphArgs, validate_scan_args, validate_graph_args};
+pub use root::RootCommand;
