+++ src/lib.rs
@@ -1,5 +1,5 @@
 pub mod analyzer;
-pub mod args;
+pub mod cli;
 pub mod call_graph;
 pub mod call_graph_output;
 pub mod config;
